pyproject.toml
[project]
name = "server-remote-sse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
